# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.EventType:
    enum:
    - wedding
    - birthday
    - corporate
    type: string
    x-enum-varnames:
    - EventTypeWedding
    - EventTypeBirthday
    - EventTypeCorporate
  models.Menu:
    properties:
      id:
        type: integer
      img:
        type: string
      name_kz:
        type: string
      name_ru:
        type: string
      restaurant_id:
        type: integer
    type: object
  models.MenuType:
    properties:
      id:
        type: integer
      img:
        type: string
      name:
        type: string
    type: object
...
  models.RestaurantEvent:
    properties:
      desc:
        type: string
      eventtype:
        $ref: '#/definitions/models.EventType'
      id:
        type: integer
      img:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API сервера для системы управления рестораном
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management System API
  version: "1.0"
paths:
...
  /events/type/{type}:
    get:
      consumes:
      - application/json
      description: Возвращает список событий ресторана указанного типа
      parameters:
      - description: Тип события (wedding, birthday, corporate)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RestaurantEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить события ресторана по типу
      tags:
      - events
  /menu-types:
    get:
      consumes:
      - application/json
      description: Возвращает список всех типов меню
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MenuType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить список всех типов меню
      tags:
      - menu-types
    post:
      consumes:
      - application/json
      description: Создает новый тип меню в системе
      parameters:
      - description: Данные типа меню
        in: body
        name: menuType
        required: true
        schema:
          $ref: '#/definitions/models.MenuType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создать новый тип меню
      tags:
      - menu-types
  /menu-types/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет тип меню по его ID
      parameters:
      - description: ID типа меню
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Удалить тип меню
      tags:
      - menu-types
    get:
      consumes:
      - application/json
      description: Возвращает тип меню по его ID
      parameters:
      - description: ID типа меню
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MenuType'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить тип меню по ID
      tags:
      - menu-types
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего типа меню
      parameters:
      - description: ID типа меню
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные типа меню
        in: body
        name: menuType
        required: true
        schema:
          $ref: '#/definitions/models.MenuType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Обновить данные типа меню
      tags:
      - menu-types
  /menus:
    post:
      consumes:
      - application/json
      description: Создает новое меню для указанного ресторана
      parameters:
      - description: Данные меню
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создать новое меню
      tags:
      - menus
  /menus/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет меню по его ID
      parameters:
      - description: ID меню
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Удалить меню
      tags:
      - menus
    get:
      consumes:
      - application/json
      description: Возвращает меню по его ID
      parameters:
      - description: ID меню
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить меню по ID
      tags:
      - menus
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего меню
      parameters:
      - description: ID меню
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные меню
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Обновить данные меню
      tags:
      - menus
  /menus/restaurant/{restaurantID}:
    get:
      consumes:
      - application/json
      description: Возвращает список меню для указанного ресторана
      parameters:
      - description: ID ресторана
        in: path
        name: restaurantID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить меню по ID ресторана
      tags:
      - menus
...
swagger: "2.0"